+++ apps/docs/src/content/04-components/content/kbd/examples/children.tsx
@@ -1,0 +1,6 @@
+import { Kbd } from "@mittwald/flow-react-components";
+
+<Kbd>
+  <Kbd keys={["mod"]} /> + <Kbd keys={["k"]} />,{" "}
+  <Kbd keys={["mod"]} /> + <Kbd keys={["c"]} />
+</Kbd>;