+++ packages/components/src/components/Kbd/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { Kbd } from "./Kbd";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    Kbd: ViewComponent<typeof Kbd>;
+  }
+}