+++ packages/components/src/components/propTypes/index.ts
@@ -83,6 +83,7 @@ import type { AlertTextProps } from "@/components/AlertText";
 import type { IllustratedMessageProps } from "@/components/IllustratedMessage";
 import type { RatingProps } from "@/components/Rating";
 import type { CodeEditorProps } from "@/components/CodeEditor";
+import type { KbdProps } from "@/components/Kbd/Kbd";
 
 export * from "./types";
 
@@ -158,6 +159,7 @@ export interface FlowComponentPropsTypes {
   Segment: SegmentProps;
   SegmentedControl: SegmentedControlProps;
   Select: SelectProps;
+  Kbd: KbdProps;
   Slider: SliderProps;
   Switch: SwitchProps;
   Tabs: TabsProps;
@@ -243,6 +245,7 @@ const propsContextSupportingComponentsMap: Record<
   Segment: true,
   SegmentedControl: true,
   Select: true,
+  Kbd: true,
   Slider: true,
   Switch: true,
   Tabs: true,