+++ packages/components/src/integrations/react-hook-form/components/Form/Form.browser.test.tsx
@@ -5,17 +5,19 @@ import TextField, { type TextFieldProps } from "@/components/TextField";
 import {
   Field,
   Form,
+  FormSettingsProvider,
   SubmitButton,
   typedField,
   useFormContext,
   type FormProps,
 } from "@/integrations/react-hook-form";
 import { Render } from "@/lib/react/components/Render";
-import type { FC } from "react";
+import type { FC, PropsWithChildren } from "react";
 import { TextArea } from "react-aria-components";
 import { useForm } from "react-hook-form";
 import { render } from "vitest-browser-react";
 import { page, userEvent } from "vitest/browser";
+import FormRootError from "../FormRootError";
 
 const handleSubmit = vitest.fn();
 vitest.useFakeTimers();
@@ -267,6 +269,54 @@ describe("readonly", () => {
   });
 });
 
+describe("FormSettingsProvider", () => {
+  const TestForm: FC<Omit<FormProps<object>, "form">> = (props) => {
+    const form = useForm<object>({
+      defaultValues: {
+        test: "value",
+      },
+    });
+    return (
+      <Form {...props} form={form}>
+        <Field name="test">
+          <TextField placeholder="textfield" aria-label="test" />
+        </Field>
+        <SubmitButton data-testid="submit-button">Submit</SubmitButton>
+      </Form>
+    );
+  };
+
+  test("form submission can be intercepted with FormSettingsProvider", async () => {
+    const errorHandler = vitest.fn();
+    window.addEventListener("unhandledrejection", errorHandler);
+
+    await render(
+      <FormSettingsProvider
+        submitInterceptor={async (submit, values) => {
+          expect(values).toEqual({ test: "value" });
+          try {
+            return await submit(values);
+          } catch {
+            // swallow error
+          }
+        }}
+      >
+        <TestForm
+          onSubmit={() => {
+            throw new Error("Submission failed");
+          }}
+        />
+      </FormSettingsProvider>,
+    );
+
+    const submitButton = page.getByTestId("submit-button");
+    await userEvent.click(submitButton);
+
+    expect(errorHandler).not.toHaveBeenCalled();
+    window.removeEventListener("unhandledrejection", errorHandler);
+  });
+});
+
 describe("error", () => {
   const TestForm: FC<Omit<FormProps<object>, "form">> = (props) => {
     const form = useForm<object>();
@@ -280,9 +330,30 @@ describe("error", () => {
     );
   };
 
+  const RootErrorTestForm: FC<PropsWithChildren> = (props) => {
+    const { children } = props;
+    const form = useForm<object>();
+    return (
+      <Form
+        {...props}
+        form={form}
+        onSubmit={() => {
+          form.setError("root", {
+            message: "Test error",
+          });
+        }}
+      >
+        <Field name="test">
+          <TextField placeholder="textfield" aria-label="test" />
+        </Field>
+        {children}
+        <SubmitButton data-testid="submit-button">Submit</SubmitButton>
+      </Form>
+    );
+  };
+
   test("form submission error leads to unhandledrejection event", async () => {
     const errorHandler = vitest.fn();
-
     window.addEventListener("unhandledrejection", errorHandler);
 
     await render(
@@ -303,6 +374,46 @@ describe("error", () => {
         }),
       }),
     );
+
+    window.removeEventListener("unhandledrejection", errorHandler);
+  });
+
+  test("form.setError('root') leads to unhandledrejection event when no FormRootError is mounted", async () => {
+    const errorHandler = vitest.fn();
+    window.addEventListener("unhandledrejection", errorHandler);
+
+    await render(<RootErrorTestForm />);
+
+    const submitButton = page.getByTestId("submit-button");
+    await userEvent.click(submitButton);
+
+    expect(errorHandler).toHaveBeenCalledWith(
+      expect.objectContaining({
+        reason: expect.objectContaining({
+          message: "Test error",
+        }),
+      }),
+    );
+
+    window.removeEventListener("unhandledrejection", errorHandler);
+  });
+
+  test("form.setError('root') leads not to unhandledrejection event when FormRootError is mounted", async () => {
+    const errorHandler = vitest.fn();
+    window.addEventListener("unhandledrejection", errorHandler);
+
+    await render(
+      <RootErrorTestForm>
+        <FormRootError />
+      </RootErrorTestForm>,
+    );
+
+    const submitButton = page.getByTestId("submit-button");
+    const errorText = page.getByText("Test error");
+    await userEvent.click(submitButton);
+
+    expect(errorHandler).not.toHaveBeenCalled();
+    expect(errorText).toBeInTheDocument();
 
     window.removeEventListener("unhandledrejection", errorHandler);
   });