+++ packages/remote-react-components/src/tests/visual/List.browser.test.tsx
@@ -32,7 +32,12 @@ test.each(testEnvironments)(
       }>();
 
       return (
-        <List.List aria-label="list" getItemId={(i) => i.id}>
+        <List.List
+          aria-label="list"
+          getItemId={(i) => i.id}
+          selectedKeys={["2"]}
+          selectionMode="multiple"
+        >
           <ActionGroup>
             <Button>Button</Button>
           </ActionGroup>
@@ -169,6 +174,8 @@ test.each(testEnvironments)(
           defaultViewMode="tiles"
           aria-label="list"
           getItemId={(i) => i.id}
+          selectedKeys={["2"]}
+          selectionMode="multiple"
         >
           <List.StaticData
             data={[
@@ -238,11 +245,20 @@ test.each(testEnvironments)(
           defaultViewMode="table"
           aria-label="list"
           getItemId={(i) => i.id}
+          selectedKeys={["1", "2"]}
+          selectionMode="multiple"
         >
           <List.StaticData
             data={[
               { id: "1", name: "Max Mustermann", role: "Admin", active: true },
               { id: "2", name: "John Doe", role: "Developer", active: false },
+              {
+                id: "3",
+                name: "John Mustermann",
+                role: "Developer",
+                active: false,
+              },
+              { id: "4", name: "Max Doe", role: "Developer", active: false },
             ]}
           />
           <List.Table>
@@ -264,5 +280,137 @@ test.each(testEnvironments)(
     await render(<Wrapper />);
 
     await testScreenshot("List table");
+  },
+);
+
+test.each(testEnvironments)(
+  "List edge cases - list view (%s)",
+  async ({
+    testScreenshot,
+    render,
+    components: {
+      typedList,
+      ListItemView,
+      Avatar,
+      Initials,
+      Heading,
+      ContextMenu,
+      Text,
+      Content,
+    },
+  }) => {
+    function Wrapper() {
+      const List = typedList<{
+        id: string;
+        name: string;
+        role: string;
+        content: string;
+        bottomContent: string;
+      }>();
+
+      return (
+        <List.List aria-label="list" getItemId={(i) => i.id}>
+          <List.StaticData
+            data={[
+              {
+                id: "1",
+                name: "MaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermannMaxMustermann",
+                role: "AdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdminAdmin",
+                content:
+                  "ContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContentContent",
+                bottomContent:
+                  "BottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContentBottomContent",
+              },
+            ]}
+          />
+          <List.Item textValue={(i) => i.name}>
+            {(i) => (
+              <ListItemView>
+                <Avatar>
+                  <Initials>{i.name}</Initials>
+                </Avatar>
+                <Heading>{i.name}</Heading>
+                <Text>{i.role}</Text>
+                <Content>{i.content}</Content>
+                <Content slot="bottom">{i.bottomContent}</Content>
+                <ContextMenu />
+              </ListItemView>
+            )}
+          </List.Item>
+        </List.List>
+      );
+    }
+
+    await render(<Wrapper />);
+
+    await testScreenshot("List edge cases - list view");
+  },
+);
+
+test.each(testEnvironments)(
+  "List edge cases - tile view (%s)",
+  async ({
+    testScreenshot,
+    render,
+    components: {
+      typedList,
+      ListItemView,
+      Avatar,
+      Initials,
+      Heading,
+      ContextMenu,
+      Text,
+      Content,
+    },
+  }) => {
+    function Wrapper() {
+      const List = typedList<{
+        id: string;
+        name: string;
+        role: string;
+        content: string;
+        bottomContent: string;
+      }>();
+
+      return (
+        <List.List
+          defaultViewMode="tiles"
+          aria-label="list"
+          getItemId={(i) => i.id}
+        >
+          <List.StaticData
+            data={[
+              {
+                id: "1",
+                name: "MaxMustermannMaxMustermann",
+                role: "AdminAdminAdminAdminAdminAdminAdmin",
+                content:
+                  "ContentContentContentContentContentContentContentContentContentContent",
+                bottomContent:
+                  "BottomContentBottomContentBottomContentBottomContentBottomContent",
+              },
+            ]}
+          />
+          <List.Item showList={false} showTiles textValue={(i) => i.name}>
+            {(i) => (
+              <ListItemView>
+                <Avatar>
+                  <Initials>{i.name}</Initials>
+                </Avatar>
+                <Heading>{i.name}</Heading>
+                <Text>{i.role}</Text>
+                <Content>{i.content}</Content>
+                <Content slot="bottom">{i.bottomContent}</Content>
+                <ContextMenu />
+              </ListItemView>
+            )}
+          </List.Item>
+        </List.List>
+      );
+    }
+
+    await render(<Wrapper />);
+
+    await testScreenshot("List edge cases - tile view");
   },
 );